Move allocator from libgc
In libgc, the user can opt-in to compiling with the rustgc fork -
which provides optimisations which need compiler support. This made
allocator linkage rather complicated: both libgc and rustgc had
implementations of the Allocator trait, which were conditionally linked
based on which features were enabled.  In addition to being a source of
confusion, this meant that allocator changes would need to be
duplicated and manually kept in-sync.

Previous attempts to link this as a separate crate failed because of a
dependency chicken-egg problem when adding it to the core library's dep
tree. However, I've recently discovered rustc-std-workspace-core[1] - a
hack that rustc uses to solve this exact issue.

[1]: https://crates.io/crates/rustc-std-workspace-core