Add specialized allocation to optimise marking
There are three ways a block can be allocated:

    Atomically - this means that a block contains no pointers which need
    tracing during marking. An atomically allocated block is a leaf node in
    the object graph.

    Precisely - blocks are allocated along with a bitmap which describes
    where the pointers are on a per-word basis. The parts of a block which
    are not described in the bitmap are ignored by the collector during
    marking.

    Conservatively - the whereabouts of pointers inside the block are
    unknown, and each word in the block must be considered for potential
    pointers.

Users who enable the rustgc feature can now allocate memory with
`GcAllocator::maybe_optimised_alloc`. This method tries to allocate
blocks precisely (or even atomically) based on a value's type.  It first
tries to specialize allocation based on whether a type implements the
`Conservative` or `NoTrace` traits, and then uses the `gc_layout`
intrinsic in rustgc to try to allocate a block with a precise layout.