Split rustgc feature into two to allow unit testing
 The rustgc feature pulls in the "core" crate. This crate is a necessary
 hack so that libgc_internal can be included in rustc-std when rustgc
 depends on it.

Unfortunately, the "core" crate prevents libgc_internal from being
compiled in any other context - even cfg(test). Thus, in order to test
the code for the "rustgc" feature, we have to de-couple conditional
compilation from the dependencies that are added by the feature gate. As
of cargo 1.50, it's not possible to exclude dependencies based on
cfg(test). A hack to get around this is to use an additional feature
(rustgc_internal) to conditionally compile rustgc-specific code only.
This way, `cargo test --features rustgc_internal` can be used to build a
test binary for rustgc-specific features, while also being automatically
included as part of the `rustgc` feature when it is compiled as part of
rustgc-std.